yolov5-object-detection/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cu113
torchvision==0.11.1+cu113
roboflow==0.2.4
Pillow